patch reflections svg
